added helper functions for standard metrics to amercian metrics, refresh button implemented last update and forgot to add
last commit had refresh button implemented on saved detail page & WORKING. current build SUCCESSFUL

+++ lib/Models/forecast_model.dart
@@ -12,7 +12,7 @@ class Forecast{
       Weather w = new Weather(
           overalls: Overall.fromJson(e["weather"][0]),
           numbers: Number.fromJson(e["main"]),
-          visibility: e["visibility"] as int,
+          visibility: e["visibility"].toDouble(),
           winds: Wind.fromJson(e["wind"]),
           currentTime: DateTime.fromMillisecondsSinceEpoch(e["dt"] * 1000, isUtc: false).toLocal(),
           systems: null,
